requirements.txt
azure-functions
azure-identity
pandas
openpyxl
pyodbc
requests_toolbelt
xlrd
pyJWT